Profiling renamed to profiling. Updated Math.java, MathTest.java. Updated GUI and Controller.java (added scientific representation of numbers).

+++ src/IMath.java
@@ -18,5 +18,5 @@ public interface IMath {
 
     double Abs(double nonAbs);
 
-    int toBinary(int basis);
+    long toBinary(long basis);
 }
